requirements.txt
Flask==0.12

# Application server for both development and production.
gunicorn==19.4

# Database.
psycopg2==2.6.1
Flask-SQLAlchemy==2.1
SQLAlchemy-Utils==0.32

# Testing and static analysis.
pytest==2.9
pytest-cov==2.2
flake8==2.5

# CLI.
Click==6.4

# API.
flask-classful==0.14
flask-marshmallow==0.7.0
flask-jwt-extended==3.3

# Extensions and libraries.
flask-debugtoolbar==0.10.0
pytz==2017.3
faker==0.8.7
pusher==1.7.2
